requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
coverage==7.2.5
Django==4.2
django-ninja==0.21.0
model-bakery==1.11.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
psycopg2-binary==2.9.6
pydantic==1.10.7
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
